Renamed mod merkle_tree to mem_backed_tree
