+++ src/lib.rs
@@ -9,7 +9,7 @@ pub mod consistency;
 pub mod error;
 pub mod inclusion;
 mod leaf;
-mod merkle_tree;
+mod mem_backed_tree;
 mod tree_math;
 
 #[cfg(test)]
@@ -17,4 +17,4 @@ mod test_util;
 
 pub use inclusion::*;
 pub use leaf::*;
-pub use merkle_tree::*;
+pub use mem_backed_tree::*;
